nvapi-vulkan: Extract factory
