+++ src/meson.build
@@ -9,6 +9,7 @@ nvapi_src = files([
   'nvapi/low_latency_frame_id_generator.cpp',
   'nvapi/nvapi_resource_factory.cpp',
   'nvapi/nvapi_d3d_low_latency_device.cpp',
+  'nvapi/nvapi_vulkan_low_latency_device_factory.cpp',
   'nvapi/nvapi_vulkan_low_latency_device.cpp',
   'nvapi/nvapi_output.cpp',
   'nvapi/nvapi_adapter.cpp',
