+++ src/nvapi/nvapi_adapter_registry.cpp
@@ -2,7 +2,7 @@
 #include "nvapi_d3d11_device.h"
 #include "nvapi_d3d12_device.h"
 #include "nvapi_d3d_low_latency_device.h"
-#include "nvapi_vulkan_low_latency_device.h"
+#include "nvapi_vulkan_low_latency_device_factory.h"
 #include "../util/util_log.h"
 
 namespace dxvk {
